% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{add_out_path_prefix}
\alias{add_out_path_prefix}
\title{Register a path prefix for the global out path}
\usage{
add_out_path_prefix(name, prefix)
}
\arguments{
\item{name}{name of the prefix}

\item{prefix}{path to add to the prefix}
}
\description{
Each path prefix will be added to the root path separated by '/' to create a full path before the suffix.
}
\details{
Prefix mechanism allows to define constant path at several levels.
For example :
output path for global out path
output path for a project (that will complete base.out.path)
output path in a script of the project
}
\examples{
\dontrun{
set_root_out_path('/my/path') # root of the out path is /my/path
add_path_prefix('project', 'my_project')
init_path('output') # Add output to the current path as suffix
}
my_path() # => '/my/path/my_project/output'
}
\seealso{
Other path-functions: 
\code{\link{create_path}()},
\code{\link{get_root_out_path}()},
\code{\link{init_path}()},
\code{\link{my_path}()},
\code{\link{print.path_components}()},
\code{\link{set_root_out_path}()}
}
\concept{path-functions}
