% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{add_path_prefix}
\alias{add_path_prefix}
\title{Register a path prefix}
\usage{
add_path_prefix(name, prefix)
}
\arguments{
\item{name}{name of the prefix}

\item{prefix}{path to add to the prefix}
}
\description{
Each path prefix will be added to `base.out.path` separated by '/' to create a full path for output
}
\details{
Prefix mechanism allows to define constant path at several levels.
For example :
output path for global workspace (base.out.path)
output path for a project (that will complete base.out.path)
output path in a script of the project
}
\examples{
# base.out.path = '/my/path/'
add_path_prefix('project', 'my_project') # A project prefix to the path
\dontrun{
init_path('output') # Add output to the current path
}
my_path() # => '/my/path/my_project/output'
}
\seealso{
Other path-functions: 
\code{\link{create_path}()},
\code{\link{get_base_out_path}()},
\code{\link{get_current_paths}()},
\code{\link{init_path}()},
\code{\link{my_path}()},
\code{\link{print.paths_definition}()},
\code{\link{set_base_out_path}()}
}
\concept{path-functions}
