% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{create_path}
\alias{create_path}
\title{Generate a path with path components without updating the builder}
\usage{
create_path(base = NULL, prefixes = list(), suffix = NULL, .builder = NULL)
}
\arguments{
\item{base}{character, base output path, use .builder will be used if NULL}

\item{prefixes}{list() replace some prefixes, only works if prefixes is already defined by \code{\link{add_out_path_prefix}()}}

\item{suffix}{character suffix to use, if NULL use default}

\item{.builder}{builder to use for default components values, if NULL use global out path builder}
}
\description{
This function build a path but do not change the current path components of the session
}
\seealso{
Other path-functions: 
\code{\link{add_out_path_prefix}()},
\code{\link{get_root_out_path}()},
\code{\link{init_path}()},
\code{\link{my_path}()},
\code{\link{print.path_components}()},
\code{\link{set_root_out_path}()}
}
\concept{path-functions}
