% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\docType{data}
\name{.out_path}
\alias{.out_path}
\title{Global out path}
\format{
An object of class \code{PathBuilder} (inherits from \code{R6}) of length 15.
}
\usage{
.out_path
}
\description{
Global out path
}
\keyword{datasets}
