% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ending_slash}
\alias{ending_slash}
\title{Ensure some paths ends with an ending slash}
\usage{
ending_slash(x)
}
\arguments{
\item{x}{path to check for ending slash}
}
\description{
Ensure some paths ends with an ending slash
}
