% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_base_out_path}
\alias{get_base_out_path}
\title{Get the global base output path}
\usage{
get_base_out_path()
}
\description{
Get the global base output path
}
\seealso{
Other path-functions: 
\code{\link{add_path_prefix}()},
\code{\link{create_path}()},
\code{\link{get_current_paths}()},
\code{\link{init_path}()},
\code{\link{my_path}()},
\code{\link{print.paths_definition}()},
\code{\link{set_base_out_path}()}
}
\concept{path-functions}
