% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_current_paths}
\alias{get_current_paths}
\title{Get the current paths defined by \code{\link{init_path}()} or \code{\link{add_path_prefix}()}}
\usage{
get_current_paths()
}
\description{
Get the current paths defined by \code{\link{init_path}()} or \code{\link{add_path_prefix}()}
}
\seealso{
Other path-functions: 
\code{\link{add_path_prefix}()},
\code{\link{create_path}()},
\code{\link{get_base_out_path}()},
\code{\link{init_path}()},
\code{\link{my_path}()},
\code{\link{print.paths_definition}()},
\code{\link{set_base_out_path}()}
}
\concept{path-functions}
