% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_root_out_path}
\alias{get_root_out_path}
\title{Get the global base output path}
\usage{
get_root_out_path()
}
\description{
Get the global base output path
}
\seealso{
Other path-functions: 
\code{\link{add_out_path_prefix}()},
\code{\link{create_path}()},
\code{\link{init_path}()},
\code{\link{my_path}()},
\code{\link{print.path_components}()},
\code{\link{set_root_out_path}()}
}
\concept{path-functions}
