% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{init_workspace}
\alias{init_workspace}
\title{Initialize a workspace}
\usage{
init_workspace(path = NULL, bootstraps = c("workspace.R"))
}
\arguments{
\item{path}{character vector, path where to init the workspace (if NULL (default), will use current directory)}

\item{bootstraps}{character vector list of file to be loaded when workspace is loaded, by default add a workpace.R file. (path must be relative to `path` argument)}
}
\description{
Create the root file .Rworkspace with list of file to load on startup
}
