% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_project.R
\name{initialise_project}
\alias{initialise_project}
\alias{initialize_project}
\title{Initialise a project}
\usage{
initialise_project(project_id)

initialize_project(project_id)
}
\arguments{
\item{project_id}{The project_id}
}
\description{
Initialise a project
}
