% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{is_out_full_path}
\alias{is_out_full_path}
\title{Test if the current output path is defined using an absolute path or has been created using internal paths segments}
\usage{
is_out_full_path()
}
\description{
Test if the current output path is defined using an absolute path or has been created using internal paths segments
}
