% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{is_workspace_booted}
\alias{is_workspace_booted}
\title{Get the booted state of the workspace}
\usage{
is_workspace_booted()
}
\description{
This function returns TRUE when workspace::launch() is called, but before the bootstrap files are loaded.
Do not use this function inside bootstrap files, because it will be always TRUE
}
