% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{is_workspace_loaded}
\alias{is_workspace_loaded}
\title{Get the loaded state of the workspace}
\usage{
is_workspace_loaded()
}
\description{
This function returns TRUE when workspace::launch() is called, after the workspace setup files are loaded.
}
