% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_R_file}
\alias{load_R_file}
\title{Load a R file by search for .R and .r version, the first found is loaded (if .R exists, the .r version is ignored)}
\usage{
load_R_file(path, must.exists = FALSE, envir = rlang::caller_env())
}
\arguments{
\item{path}{file path (without extension) to find}

\item{must.exists}{if TRUE, raise an error if the file is not find with .R or .r extension}

\item{envir}{environment where to load the file, by default the caller environment}
}
\description{
Load a R file by search for .R and .r version, the first found is loaded (if .R exists, the .r version is ignored)
}
