% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_workspace}
\alias{load_workspace}
\title{Load Workspace bootstrap files}
\usage{
load_workspace(envir = rlang::caller_env())
}
\arguments{
\item{envir}{Environment to load the bootstrap files into, by default the caller env}
}
\description{
Find the workspace location and load files listed in .Rworkspace in the passed environment
}
