% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{my_path}
\alias{my_path}
\title{Return the path of a file in the global out path}
\usage{
my_path(...)
}
\arguments{
\item{...}{characters string to used (will be concatenated with no space)}
}
\description{
my_path is the common function to create path inside the global out path.
}
\examples{
\dontrun{
  set_root_out_path("/my/path")
  init_path("")
  my_path("toto") # get "toto" inside the out path -> /my/path/toto
}
}
\seealso{
Other path-functions: 
\code{\link{add_out_path_prefix}()},
\code{\link{create_path}()},
\code{\link{get_root_out_path}()},
\code{\link{init_path}()},
\code{\link{print.path_components}()},
\code{\link{set_root_out_path}()}
}
\concept{path-functions}
