% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{new_workspace}
\alias{new_workspace}
\title{Create a new workspace}
\usage{
new_workspace(dir = tempfile(pattern = "ws"))
}
\arguments{
\item{dir}{directory used to store datasets, by default uses temporary directory.}
}
\value{
an empty workspace object.
}
\description{
Create a new workspace, i.e. an object made of an
R environment and a directory where datasets will be
stored as parquet files.
}
\examples{
z <- new_workspace()
z
}
\seealso{
\link{workspace} for package documentation

Other functions to manage workspaces:
\code{\link[=pack_workspace]{pack_workspace()}},
\code{\link[=unpack_workspace]{unpack_workspace()}},
\code{\link[=workspace_bind]{workspace_bind()}},
\code{\link[=workspace_copy]{workspace_copy()}}
}
\concept{functions to manage workspaces}
