% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack-folder.R
\name{pack_folder}
\alias{pack_folder}
\title{Compress a folder}
\usage{
pack_folder(folder, target)
}
\arguments{
\item{folder}{folder to compress}

\item{target}{path of the archive to create}
}
\value{
Returns path to \code{target}.
}
\description{
Compress a folder to a target file. The
function returns the complete path to target file.
}
