% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{print.path_components}
\alias{print.path_components}
\title{Print path components}
\usage{
\method{print}{path_components}(x, ...)
}
\arguments{
\item{x}{paths_definitions object}

\item{...}{extra params (print interface)}
}
\description{
Print the path components returns by the components method of a path builder
}
\seealso{
Other path-functions: 
\code{\link{add_out_path_prefix}()},
\code{\link{create_path}()},
\code{\link{get_root_out_path}()},
\code{\link{init_path}()},
\code{\link{my_path}()},
\code{\link{set_root_out_path}()}
}
\concept{path-functions}
