% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{print.paths_definition}
\alias{print.paths_definition}
\title{Print path definition}
\usage{
\method{print}{paths_definition}(x, ...)
}
\arguments{
\item{x}{paths_definitions object}

\item{...}{extra params (print interface)}
}
\description{
Print the path definitions return by \code{\link{get_current_paths}}
}
\seealso{
Other path-functions: 
\code{\link{add_path_prefix}()},
\code{\link{create_path}()},
\code{\link{get_base_out_path}()},
\code{\link{get_current_paths}()},
\code{\link{init_path}()},
\code{\link{my_path}()},
\code{\link{set_base_out_path}()}
}
\concept{path-functions}
