% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R, R/project_subfolders.R
\name{project}
\alias{project}
\alias{root_file}
\alias{script_file}
\alias{derived_file}
\alias{raw_file}
\alias{result_file}
\title{Helper function for controlling projects}
\usage{
project(project_id, experiment_id = NULL)

root_file(filename = "", project_id = NULL, experiment_id = NULL)

script_file(filename = "", project_id = NULL, experiment_id = NULL)

derived_file(filename = "", project_id = NULL, experiment_id = NULL)

raw_file(filename = "", project_id = NULL, experiment_id = NULL)

result_file(filename = "", project_id = NULL, experiment_id = NULL)
}
\arguments{
\item{project_id}{id for the project}

\item{filename}{the filename}
}
\description{
Helper function for controlling projects
}
\examples{
\dontrun{
project("test_plots")

data <- matrix(runif(200), ncol = 2)
pdf(figure_file("testplot.pdf"), 5, 5)
plot(data)
dev.off()
}
}
