% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{set_base_out_path}
\alias{set_base_out_path}
\title{Define the global base output path}
\usage{
set_base_out_path(path)
}
\arguments{
\item{path}{path where the output path should be defined}
}
\description{
This function should be called before any use of \code{\link{init_path}()} or \code{\link{my_path}()}
It only defines the global output path. It's intented to be defined on startup and not changed during the script execution.
If it's has to be changed, then \code{\link{init_path}()} should be called after to initialize the current output path.
}
\seealso{
Other path-functions: 
\code{\link{add_path_prefix}()},
\code{\link{create_path}()},
\code{\link{get_base_out_path}()},
\code{\link{get_current_paths}()},
\code{\link{init_path}()},
\code{\link{my_path}()},
\code{\link{print.paths_definition}()}
}
\concept{path-functions}
