% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace-package.R
\docType{package}
\name{workspace-package}
\alias{workspace}
\alias{workspace-package}
\title{workspace: Effortlessly Create and Restore Development Environments}
\description{
The {workspace} package aims to provide a seamless experience in creating and restoring development environments effortlessly. It offers efficient tools and functionalities to manage workspace state, allowing users to easily save and revert to previous setups. The package facilitates improved productivity and organization for developers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KoderKow/workspace}
  \item \url{https://koderkow.github.io/workspace/}
  \item Report bugs at \url{https://github.com/KoderKow/workspace/issues}
}

}
\author{
\strong{Maintainer}: Kyle Harris \email{koderkow@gmail.com} [copyright holder]

}
\keyword{internal}
