% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{workspace_options}
\alias{workspace_options}
\title{Define state of the package}
\usage{
workspace_options(...)
}
\arguments{
\item{...}{list of named argument with options to defined}
}
\description{
Define state of the package
}
\details{
Accepted options
\describe{
 \itemize{base.out.path}{}
}
}
