% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_save.R
\name{workspace_save}
\alias{workspace_save}
\title{Create a restore point for the current dev environment}
\usage{
workspace_save(note = "", n_threads = 2, is_interactive = interactive())
}
\arguments{
\item{note}{Character. Default \code{""}. An optional note to add to the restore point, providing additional context or information.}

\item{n_threads}{Integer. Default \code{2}. The number of threads to use for the saving process.}

\item{is_interactive}{Logical. Default \code{interactive()}. This parameter is used for package development purposes only, and it is recommended to avoid changing this setting.}
}
\value{
Nothing.
}
\description{
This function creates a restore point for the current development environment, allowing you to save the state of your workspace for future restoration.
}
\seealso{
Other workspace functions: 
\code{\link{workspace_restore}()}
}
\concept{workspace functions}
