% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspace_state}
\alias{workspace_state}
\title{Define state of the package}
\usage{
workspace_state(...)
}
\arguments{
\item{...}{named list of values to set in the state}
}
\value{
environment the current state
}
\description{
Define state of the package
}
