% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_workspace}
\alias{find_workspace}
\title{Find workspace root}
\usage{
find_workspace(max.depth = 20)
}
\arguments{
\item{max.depth}{maximum number of upper directory to scan. Default is reasonable but it can be changed.}
}
\description{
The root of the workspace is done by finding location of the .Rworkspace file
This function can be called in subdirectories of a project to find the workspace root and be able to load
The workspace configuration.
}
