% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{my_path}
\alias{my_path}
\title{Return the path of a file in the current ouput path}
\usage{
my_path(...)
}
\arguments{
\item{...}{characters string to used (will be concatenated with no space)}
}
\description{
Return the path of a file in the current ouput path
}
\seealso{
Other path-functions: 
\code{\link{add_path_prefix}()},
\code{\link{create_path}()},
\code{\link{get_base_out_path}()},
\code{\link{get_current_paths}()},
\code{\link{init_path}()},
\code{\link{print.paths_definition}()},
\code{\link{set_base_out_path}()}
}
\concept{path-functions}
