% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{new_workspace}
\alias{new_workspace}
\title{Create a new workspace}
\usage{
new_workspace(dir = tempfile(pattern = "ws"))
}
\arguments{
\item{dir}{directory used to store datasets}
}
\description{
Create a new workspace, i.e. an object made of an
R environment and a directory where datasets will be
stored as parquet files.
}
\examples{
z <- new_workspace()
z
}
