% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{store_dataset}
\alias{store_dataset}
\title{Store a dataset into a workspace}
\usage{
store_dataset(
  x,
  dataset,
  name,
  timestamp = format(Sys.time(), "\%Y-\%m-\%d \%H:\%M:\%S")
)
}
\arguments{
\item{x}{the workspace object}

\item{dataset}{the data.frame to store in the workspace}

\item{name}{name associated with the data.frame, if an workspace file with this name exists
already it will be replaced.}

\item{timestamp}{A timestamp string to associate with the entry in the workspace.}
}
\description{
Store a dataset as a parquet file into an existing workspace.
}
\examples{
library(workspace)
dir_tmp <- tempfile(pattern = "ws")
z <- new_workspace(dir = dir_tmp)
z <- store_dataset(x = z, dataset = iris, name = "iris_dataset")
z <- store_dataset(x = z, dataset = mtcars, name = "mtcars")
z
}
\seealso{
Other functions to write in a workspace: 
\code{\link{delete_dataset}()},
\code{\link{store_json}()},
\code{\link{store_rds}()}
}
\concept{functions to write in a workspace}
