% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace-package.R
\docType{package}
\name{workspace-package}
\alias{workspace}
\alias{workspace-package}
\title{workspace: Effortlessly create and restore dev environments}
\description{
The goal of {workspace} is to effortlessly create and restore dev environments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KoderKow/workspace}
  \item Report bugs at \url{https://github.com/KoderKow/workspace/issues}
}

}
\author{
\strong{Maintainer}: Kyle Harris \email{koderkow@gmail.com}

}
\keyword{internal}
