% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.R
\docType{package}
\name{workspace-package}
\alias{workspace}
\alias{workspace-package}
\title{Workspace Package}
\description{
This package provides a set of functions to manage a group of R scripts sharing a common configuration.
}
\details{

}
\section{Options}{


Several options can be defined in options()

\describe{
\item{workspace.verbose}{Show verbose message during workspace loading}
\item{workspace.outpath}{Define base output path (call \code{\link{set_root_out_path}()})}
\item{workspace.autoload}{Autoload workspace when library is loaded}
\item{workspace.env}{Environment where to load workspace bootstap files when using autoload, default is globalenv}
\item{workspace.maxdepth}{Max depth to use to find the workspace accross upper directory of getwd(), default is 100}
}
}

\author{
\strong{Maintainer}: Clément Turbelin \email{clement.turbelin@sorbonne-universite.f}

}
