% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_restore.R
\name{workspace_restore}
\alias{workspace_restore}
\title{Restore the dev environment from a restore point}
\usage{
workspace_restore(n_threads = 2, is_interactive = interactive())
}
\arguments{
\item{n_threads}{Integer. Default \code{2}. The number of threads to use for the saving process.}

\item{is_interactive}{Logical. Default \code{interactive()}. This parameter is used for package development purposes only, and it is recommended to avoid changing this setting.}
}
\value{
Nothing.
}
\description{
This function restores the development environment from a previously created restore point, allowing you to revert your workspace to a previous state.
}
\seealso{
Other workspace functions: 
\code{\link{workspace_save}()}
}
\concept{workspace functions}
